Created CombatComponent to handle combat related functionality
BlasterCharacter is a friend class of the combat component to access
all protected and private members on  the CombatComponent.

I added an equip weapon function to the combat component which only
happens on the server so the server is in charge of equipping weapons.

Clients are currently unable to equip weapons.